requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
Django==3.2.6
Fiona==1.8.20
geopandas==0.9.0
munch==2.5.0
numpy==1.21.1
opencv-python==4.5.3.56
pandas==1.3.1
Pillow==8.3.1
pyproj==3.1.0
python-dateutil==2.8.2
pytz==2021.1
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.1
